requirements.txt
# Web framework
fastapi
uvicorn

# Database
psycopg2-binary
asyncpg

# Embeddings
voyageai

# Model inference
torch
numpy

# Utilities
python-dotenv